+++ tests/options/pugAttributeSeparator/none/none.test.ts
@@ -27,7 +27,8 @@ describe('Options', () => {
       expect(actual).toBe(expected);
       expect(() => {
         // ts-jest needs the exclamation mark, so it does not have the impression that the variables are undefined
-        /* eslint @typescript-eslint/no-non-null-assertion: off */
+        /* eslint-disable @typescript-eslint/no-non-null-assertion */
+        /* eslint-disable @typescript-eslint/no-unnecessary-type-assertion */
         parsers!.pug!.parse(actual, parsers!, null!);
       }).toThrow('Assigning to rvalue');
     });